rename in memory package and services closes #93
