+++ pet-clinic-data/src/main/java/com/daddyrusher/petclinic/service/inmemory/InMemoryPetTypeService.java
@@ -1,4 +1,4 @@
-package com.daddyrusher.petclinic.service.map;
+package com.daddyrusher.petclinic.service.inmemory;
 
 import com.daddyrusher.petclinic.model.PetType;
 import com.daddyrusher.petclinic.service.PetTypeService;
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import java.util.Set;
 
 @Service
-public class PetTypeServiceMap extends AbstractMapService<PetType, Long> implements PetTypeService {
+public class InMemoryPetTypeService extends AbstractInMemoryService<PetType, Long> implements PetTypeService {
     @Override
     public Set<PetType> findAll() {
         return super.findAll();
