+++ pet-clinic-data/src/main/java/com/daddyrusher/petclinic/service/inmemory/InMemorySpecialtyService.java
@@ -1,4 +1,4 @@
-package com.daddyrusher.petclinic.service.map;
+package com.daddyrusher.petclinic.service.inmemory;
 
 import com.daddyrusher.petclinic.model.Specialty;
 import com.daddyrusher.petclinic.service.SpecialtyService;
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import java.util.Set;
 
 @Service
-public class SpecialtyServiceMap extends AbstractMapService<Specialty, Long> implements SpecialtyService {
+public class InMemorySpecialtyService extends AbstractInMemoryService<Specialty, Long> implements SpecialtyService {
     @Override
     public Set<Specialty> findAll() {
         return super.findAll();
